Score now works
Player score will update when eating food pellets.

Note: we should work on getting seperate scenes for main menu and end
game screens. Score and who won needs to be shown at the end of the
game. Current placeholder names for scores are "Player 1: value" and
"Player 2: value". should be replaced with the username when they sign
into game later